pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "signaliq"
version = "0.1.0"
description = "Real-time vision intelligence layer for B2B sales calls"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "opencv-python-headless>=4.8.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "aiosqlite>=0.19.0",
    "websockets>=12.0",
    "pydantic>=2.0",
    "numpy>=1.24.0",
    "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
ml = [
    "fer>=22.5.1",
    "deepface>=0.0.90",
    "mediapipe>=0.10.0",
    "tensorflow>=2.16.0,<2.20.0",
]
agents = [
    "vision-agents[getstream,gemini,elevenlabs,deepgram,ultralytics]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
include = ["agent*", "api*"]
